feat(forwarder): /api/v2/timeseries streams=events now emits classified rows (#466)
## Summary

- `/api/v2/timeseries?streams=events` previously returned **501 Not
Implemented**. This change wires the existing kind/priority taxonomy SQL
onto the SSE pipeline so the live + archive event stream works
end-to-end.
- Extracted the ~430-line taxonomy SQL out of `main.go` into a
single-source-of-truth helper (`events_query.go`) so both
`/api/v2/session_events` (used by the v3 dashboard's archive events
lane) and the new timeseries path compute kind/priority identically.
- Added a 2s polling loop (`events_stream.go`) that re-runs the SQL
against `from=highWaterMark` while the SSE connection is open. New
events past the last-seen ts surface within one tick.

## What changed

| File | Change |
|---|---|
| `events_query.go` (new) | Exports `buildEventsQuery`,
`runEventsQuery`, `eventsSQLTemplate`. |
| `events_stream.go` (new) | `emitBackfillEvents` (one-shot) +
`startEventsPoller` (background ticker, cancel-on-ctx). |
| `timeseries.go` | Removed 501 stub; added `case streamEvents` in
`emitBackfill`; threaded `*sync.Mutex` through `streamLiveDeltas` so
frames from the poller and ring loop don't interleave. |
| `main.go` | `sessionEventsHandler` now calls `buildEventsQuery` —
duplicate inline SQL deleted (~430 lines). |
| `Dockerfile` | `go work edit -dropuse ./tools/harness-cli` so the
forwarder build doesn't need the CLI module that sits in the dev host's
go.work for editor convenience. |
| `streambundles.go` | Comment fix: three stale "v3 endpoint/design"
refs → "v2 timeseries". |

Net: +635 / −440. The deletion side is the inline-SQL duplicate.

## Design notes

- **Live derivation is intentionally not reimplemented in Go.** The
taxonomy is non-trivial (stall pairing via leadInFrame, lag-based
counter detection per play). Reimplementing it would fork the
classification from what the dashboard archive queries return — a
guaranteed source of drift bugs. Polling CH every 2s keeps the
kind/priority logic in exactly one place.
- **2s cadence** matches the dashboard's archive-refresh interval.
Tighter (1s) adds CH load without changing operator perception; slacker
(5s+) makes "fault add → fault_on event" feel laggy.
- **Dedupe** uses `ts|type|info` because CH ts has ms precision and
multiple events can land in the same tick (e.g. `fault_on` +
`transport_failure` from the same snapshot).

## Test plan

- [x] `go build ./...` clean in `analytics/go-forwarder/`
- [x] `go vet ./...` clean
- [x] Forwarder image builds (`docker compose build forwarder`) — the
Dockerfile change is required
- [x] `/api/v2/session_events` returns byte-identical JSON to before
(legacy dashboard path unchanged)
- [x] `curl /analytics/api/v2/timeseries?streams=events&bundles=events`
returns SSE event frames (was 501)
- [x] Verified live + backfill against the test-dev iPad session: 2,625
classified events spanning the full taxonomy (stall, downshift, upshift,
buffering, restart, loop_server, http_5xx, slow_request, request_retry,
fault_on, all_failure, etc.)
- [x] Live updates visible during open subscription (restart → buffering
→ playback_start cluster appeared within ~2s of underlying snapshots)
- [ ] Dashboard's events lane still renders correctly (archive query is
byte-identical so this should pass without changes)
