+++ analytics/go-forwarder/main.go
@@ -1221,59 +1221,6 @@ func serveHTTP(ctx context.Context, cfg config, ring *Ring) {
 			return
 		}
 
-		params := map[string]string{}
-		clauses := []string{}
-		if session != "" {
-			clauses = append(clauses, "session_id = {session:String}")
-			params["session"] = session
-		}
-		if playerID != "" {
-			// Case-insensitive: device-side player_ids often arrive in
-			// the case the player reported (often uppercase) while v2
-			// /api/v2/players normalises to lowercase. lowerUTF8 on both
-			// sides papers over the mismatch.
-			clauses = append(clauses, "lowerUTF8(player_id) = lowerUTF8({player:String})")
-			params["player"] = playerID
-		}
-		// HAR events come from network_requests, which on iOS often
-		// has play_id='' on variant/segment URLs (the iOS player
-		// strips the query param). Filter HAR by time range derived
-		// from session_snapshots instead of play_id directly.
-		// We qualify ts as `nr.ts` because the inner sub-selects
-		// alias `toString(ts) AS ts` and that String alias would
-		// otherwise shadow the column in the WHERE clause.
-		harClauses := append([]string{}, clauses...)
-		if v := r.URL.Query().Get("play_id"); v != "" {
-			var pidPred string
-			if v == "—" {
-				pidPred = "play_id = ''"
-			} else {
-				pidPred = "lowerUTF8(play_id) = lowerUTF8({play:String})"
-				params["play"] = v
-			}
-			clauses = append(clauses, pidPred)
-			// Inner subqueries need an identifying WHERE that mirrors
-			// whichever id(s) the caller supplied.
-			idWhere := []string{}
-			if session != "" {
-				idWhere = append(idWhere, "session_id = {session:String}")
-			}
-			if playerID != "" {
-				idWhere = append(idWhere, "lowerUTF8(player_id) = lowerUTF8({player:String})")
-			}
-			idClause := strings.Join(idWhere, " AND ")
-			harClauses = append(harClauses, fmt.Sprintf(
-				"nr.ts BETWEEN (SELECT min(ts) FROM %s.%s WHERE %s AND %s) "+
-					"AND (SELECT max(ts) FROM %s.%s WHERE %s AND %s)",
-				cfg.chDatabase, cfg.chTable, idClause, pidPred,
-				cfg.chDatabase, cfg.chTable, idClause, pidPred))
-		}
-		// Default 5000 is enough for a quick smoke test; cap 50000
-		// accommodates long sessions where ABR upshift / downshift
-		// noise (often > 60% of events) crowds out rare fault events
-		// at smaller LIMITs. The DESC sort keeps the most-recent N
-		// regardless, so raising the cap is "show more history" not
-		// "fix correctness".
 		limit := 5000
 		if v := r.URL.Query().Get("limit"); v != "" {
 			var n int
@@ -1281,371 +1228,20 @@ func serveHTTP(ctx context.Context, cfg config, ring *Ring) {
 				limit = n
 			}
 		}
-		where := strings.Join(clauses, " AND ")
-		harWhere := strings.Join(harClauses, " AND ")
-		// Event taxonomy:
-		//   - Player-emitted events come from the last_event column directly
-		//     (the player tells us "buffering_start", "rate_shift_down", etc.)
-		//   - Stall vs buffering are kept distinct: AVPlayer (iOS) emits both
-		//     and they mean different things; Chrome/HLS.js only emits stall_*
-		//   - Stall and buffering durations are computed by pairing onsets
-		//     with the next end event in the same family (stall vs buffer)
-		//   - Server-side counters (faults, manifest/segment failures) are
-		//     not surfaced by the player, so we keep lagInFrame for those
-		//
-		// Priority (computed in outer SELECT):
-		//   P1 Critical : error, master/all failure, stall ≥ 3s
-		//   P2 High     : stall < 3s, restart, manifest/segment/transport_failure,
-		//                 transfer_*_timeout
-		//   P3 Medium   : downshift, fault_on, timejump, buffering
-		//   P4 Low      : upshift, fault_off, playback_start
-		query := fmt.Sprintf(`
-			WITH stall_or_buffer_pairs AS (
-			  SELECT start_ts, start_event, duration_s
-			  FROM (
-			    SELECT
-			      ts AS start_ts,
-			      last_event AS start_event,
-			      multiIf(
-			        last_event IN ('stall_start','stall_end'),         'stall',
-			        last_event IN ('buffering_start','buffering_end'), 'buffer',
-			        ''
-			      ) AS family,
-			      leadInFrame(last_event, 1, '') OVER w AS next_event,
-			      dateDiff('millisecond', ts, leadInFrame(ts, 1, ts) OVER w) / 1000.0 AS duration_s
-			    FROM %s.%s
-			    WHERE %s
-			      AND last_event IN ('stall_start','stall_end','buffering_start','buffering_end')
-			    WINDOW w AS (PARTITION BY family ORDER BY ts
-			                 ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING)
-			  )
-			  -- Only emit when the lead landed on the matching close event;
-			  -- otherwise we'd surface the gap to the NEXT start (which can
-			  -- be hours away — happens when an _end was never received).
-			  WHERE (start_event = 'stall_start'     AND next_event = 'stall_end')
-			     OR (start_event = 'buffering_start' AND next_event = 'buffering_end')
-			),
-			rate_shifts AS (
-			  SELECT ts, last_event, video_bitrate_mbps,
-			         lagInFrame(video_bitrate_mbps, 1, video_bitrate_mbps) OVER w AS prev_bitrate
-			  FROM %s.%s
-			  WHERE %s
-			  WINDOW w AS (ORDER BY ts)
-			),
-			base AS (
-			  SELECT
-			    ts,
-			    player_error, transport_fault_active,
-			    manifest_consecutive_failures,
-			    segment_consecutive_failures,
-			    master_manifest_consecutive_failures,
-			    all_consecutive_failures,
-			    transport_consecutive_failures,
-			    fault_count_transfer_active_timeout,
-			    fault_count_transfer_idle_timeout,
-			    loop_count_server,
-			    -- row_number used to suppress first-snapshot artifacts: the
-			    -- first row per play has no real prior row, so the lag-default
-			    -- would otherwise spurious-fire if a counter is non-zero on
-			    -- entry (e.g. transport_consecutive_failures = 1 from a prior
-			    -- session leaving state behind).
-			    row_number() OVER w AS rn,
-			    lagInFrame(player_error, 1, '')                       OVER w AS prev_error,
-			    lagInFrame(transport_fault_active, 1, 0)              OVER w AS prev_fault,
-			    lagInFrame(manifest_consecutive_failures, 1, 0)       OVER w AS prev_manifest_fail,
-			    lagInFrame(segment_consecutive_failures, 1, 0)        OVER w AS prev_segment_fail,
-			    lagInFrame(master_manifest_consecutive_failures, 1, 0) OVER w AS prev_master_fail,
-			    lagInFrame(all_consecutive_failures, 1, 0)            OVER w AS prev_all_fail,
-			    lagInFrame(transport_consecutive_failures, 1, 0)      OVER w AS prev_transport_fail,
-			    lagInFrame(fault_count_transfer_active_timeout, 1, 0) OVER w AS prev_active_to,
-			    lagInFrame(fault_count_transfer_idle_timeout, 1, 0)   OVER w AS prev_idle_to,
-			    lagInFrame(loop_count_server, 1, 0)                   OVER w AS prev_loop_server
-			  FROM %s.%s
-			  WHERE %s
-			  WINDOW w AS (PARTITION BY play_id ORDER BY ts)
-			)
-			SELECT
-			  ts, type, info,
-			  -- 'cause' = proxy / system action that *might* produce a user-visible
-			  -- effect (fault injection, server-side failure counters, timeouts).
-			  -- 'effect' = something the player or user actually saw (stall, error,
-			  -- bitrate change, restart). Causes default to P3 (diagnostic context)
-			  -- unless they are the kill-switch types that by design break playback.
-			  multiIf(
-			    -- Inverted classification: explicitly enumerate KNOWN
-			    -- causes (proxy/system actions); everything else —
-			    -- including unrecognised player-emitted events from
-			    -- the catch-all UNION ALL — defaults to 'effect'.
-			    type IN ('master_manifest_failure', 'all_failure',
-			             'manifest_failure', 'segment_failure',
-			             'transport_failure',
-			             'transfer_active_timeout', 'transfer_idle_timeout',
-			             'fault_on', 'fault_off',
-			             'http_5xx', 'http_4xx',
-			             'request_timeout', 'request_incomplete', 'request_faulted',
-			             'slow_request', 'slow_segment', 'request_retry',
-			             'loop_server'), 'cause',
-			    'effect'
-			  ) AS kind,
-			  multiIf(
-			    -- User-marked moments are explicit "this matters" flags
-			    -- from a human watching live — always P1 so they don't
-			    -- get hidden by the default Critical+High+Medium chip
-			    -- selection.
-			    type = 'user_marked', 1,
-			    type = 'error', 1,
-			    -- These two causes break playback by design — stay Critical.
-			    type IN ('master_manifest_failure', 'all_failure'), 1,
-			    type = 'stall' AND duration_s >= 3, 1,
-			    type = 'stall', 2,
-			    type = 'restart', 2,
-			    -- Other causes are diagnostic context, not Critical/High by default.
-			    type IN ('manifest_failure', 'segment_failure',
-			             'transport_failure',
-			             'transfer_active_timeout', 'transfer_idle_timeout',
-			             'fault_on', 'fault_off'), 3,
-			    type IN ('downshift', 'timejump', 'buffering'), 3,
-			    -- HAR-derived events. http_5xx & request_timeout are
-			    -- High because they indicate real failures that often
-			    -- precede a stall. The rest are diagnostic Medium /
-			    -- Low (retries especially are very chatty).
-			    type IN ('http_5xx', 'request_timeout'), 2,
-			    type IN ('http_4xx', 'request_incomplete', 'request_faulted',
-			             'slow_request', 'slow_segment'), 3,
-			    type = 'request_retry', 4,
-			    type IN ('upshift', 'playback_start'), 4,
-			    -- Server loop boundaries are routine in this testing
-			    -- platform (the stream loops the source repeatedly).
-			    -- P4 keeps them available for triage but hidden by
-			    -- default so they don't drown out real signals.
-			    type = 'loop_server', 4,
-			    -- Default for unrecognised types: Medium so they're
-			    -- visible (vs P4 Low which is off by default).
-			    3
-			  ) AS priority
-			FROM (
-			  -- Stalls (paired stall_start → stall_end)
-			  SELECT toString(start_ts) AS ts, 'stall' AS type,
-			         concat(toString(round(duration_s, 2)), 's') AS info,
-			         duration_s
-			  FROM stall_or_buffer_pairs
-			  WHERE start_event = 'stall_start' AND duration_s > 0
-
-			  UNION ALL
-			  -- Buffering (paired buffering_start → buffering_end). Distinct
-			  -- from stall: AVPlayer emits both as separate notifications.
-			  SELECT toString(start_ts) AS ts, 'buffering' AS type,
-			         concat(toString(round(duration_s, 2)), 's') AS info,
-			         duration_s
-			  FROM stall_or_buffer_pairs
-			  WHERE start_event = 'buffering_start' AND duration_s > 0
-
-			  UNION ALL
-			  -- frozen / segment_stall: no explicit *_end emitted, so duration
-			  -- is unknown — emit as zero-duration stall events. frozen is
-			  -- always Critical via a hard-coded subtype hint in info.
-			  SELECT toString(ts) AS ts, 'stall' AS type,
-			         if(last_event = 'frozen', '(frozen)', '(segment)') AS info,
-			         0 AS duration_s
-			  FROM %s.%s WHERE %s AND last_event IN ('frozen', 'segment_stall')
-
-			  UNION ALL
-			  -- Restart, playback_start, downshift, upshift, timejump, error
-			  -- (player-emitted directly via last_event)
-			  SELECT toString(ts) AS ts, 'restart' AS type, '' AS info, 0 AS duration_s
-			  FROM %s.%s WHERE %s AND last_event = 'restart'
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'playback_start' AS type, '' AS info, 0
-			  FROM %s.%s WHERE %s AND last_event IN ('video_start_time', 'video_first_frame')
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'downshift' AS type,
-			         concat(toString(round(prev_bitrate, 2)), '→', toString(round(video_bitrate_mbps, 2)), ' Mbps') AS info,
-			         0
-			  FROM rate_shifts WHERE last_event = 'rate_shift_down' AND prev_bitrate > 0 AND video_bitrate_mbps > 0
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'upshift' AS type,
-			         concat(toString(round(prev_bitrate, 2)), '→', toString(round(video_bitrate_mbps, 2)), ' Mbps') AS info,
-			         0
-			  FROM rate_shifts WHERE last_event = 'rate_shift_up' AND prev_bitrate > 0 AND video_bitrate_mbps > 0
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'timejump' AS type, '' AS info, 0
-			  FROM %s.%s WHERE %s AND last_event = 'timejump'
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'error' AS type, player_error AS info, 0
-			  FROM %s.%s WHERE %s AND last_event = 'error'
-			  UNION ALL
-			  -- User-marked moments — the iOS / iPadOS player has a
-			  -- "911" button that fires last_event='user_marked' so an
-			  -- operator viewing live can flag "something interesting
-			  -- just happened" without us having to detect what. Always
-			  -- surface as Critical so it's never hidden by chip filter.
-			  SELECT toString(ts) AS ts, 'user_marked' AS type, '' AS info, 0
-			  FROM %s.%s WHERE %s AND last_event = 'user_marked'
-			  UNION ALL
-			  -- Catch-all: any last_event value we haven't explicitly
-			  -- mapped above shows up here verbatim. Lets new player
-			  -- events appear in the dropdown without a forwarder
-			  -- redeploy. 'heartbeat' / 'state_change' / 'playing' /
-			  -- 'video_bitrate_change' are intentionally excluded as
-			  -- noise. The explicitly-handled types are also excluded
-			  -- so we don't double-emit.
-			  SELECT toString(ts) AS ts, last_event AS type, '' AS info, 0
-			  FROM %s.%s WHERE %s
-			    AND last_event != ''
-			    AND last_event NOT IN (
-			      'heartbeat', 'state_change', 'playing', 'video_bitrate_change',
-			      'stall_start', 'stall_end',
-			      'buffering_start', 'buffering_end',
-			      'frozen', 'segment_stall',
-			      'restart', 'video_first_frame', 'video_start_time',
-			      'rate_shift_down', 'rate_shift_up',
-			      'timejump', 'error', 'user_marked'
-			    )
-
-			  -- Server-side state-change events (lag-based — player can't
-			  -- tell us about these). All filters require rn > 1 so the
-			  -- first snapshot per play (where the lag-default of 0 / ''
-			  -- would falsely trigger if the counter started non-zero) is
-			  -- suppressed.
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'error' AS type, player_error AS info, 0
-			  FROM base WHERE rn > 1 AND player_error != '' AND prev_error != player_error
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'master_manifest_failure' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND master_manifest_consecutive_failures > prev_master_fail
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'all_failure' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND all_consecutive_failures > prev_all_fail
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'manifest_failure' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND manifest_consecutive_failures > prev_manifest_fail
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'segment_failure' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND segment_consecutive_failures > prev_segment_fail
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'transport_failure' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND transport_consecutive_failures > prev_transport_fail
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'transfer_active_timeout' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND fault_count_transfer_active_timeout > prev_active_to
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'transfer_idle_timeout' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND fault_count_transfer_idle_timeout > prev_idle_to
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'fault_on' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND transport_fault_active = 1 AND prev_fault = 0
-			  UNION ALL
-			  SELECT toString(ts) AS ts, 'fault_off' AS type, '' AS info, 0
-			  FROM base WHERE rn > 1 AND transport_fault_active = 0 AND prev_fault = 1
-			  UNION ALL
-			  -- Server-loop boundary: loop_count_server increments each
-			  -- time the live LL-HLS/DASH worker rotates back to loop 0
-			  -- of the source. Useful as a periodic gridline for long
-			  -- testing sessions ("did the stall straddle a loop?").
-			  SELECT toString(ts) AS ts, 'loop_server' AS type,
-			         concat('loop ', toString(loop_count_server)) AS info, 0
-			  FROM base WHERE rn > 1 AND loop_count_server > prev_loop_server
-
-			  -- Network-request-derived events. These come from the
-			  -- network_requests table (HAR archive). Mirrors the
-			  -- waterfall's noteworthy classes: status-4xx/5xx,
-			  -- fault-timeout, fault-incomplete, is-retry, slow-transfer.
-			  -- Each row produces exactly one event via mutually-
-			  -- exclusive WHEREs; nothing double-counts.
-			  -- HAR sub-selects use a table alias 'nr' so the harWhere
-			  -- clause's nr.ts BETWEEN ... resolves to the DateTime64
-			  -- column. Without it, the projection's toString(ts) AS ts
-			  -- would shadow ts as a String and break the comparison.
-			  UNION ALL
-			  SELECT toString(nr.ts) AS ts, 'http_5xx' AS type,
-			         concat(toString(status), ' ', method, ' ', path) AS info, 0 AS duration_s
-			  FROM %s.network_requests AS nr WHERE %s AND status >= 500
-			  UNION ALL
-			  SELECT toString(nr.ts) AS ts, 'http_4xx' AS type,
-			         concat(toString(status), ' ', method, ' ', path) AS info, 0
-			  FROM %s.network_requests AS nr WHERE %s AND status >= 400 AND status < 500
-			  UNION ALL
-			  -- Faulted requests, categorised by fault_type:
-			  --   * 'timeout'  → request_timeout   (P2: picture may stop)
-			  --   * corrupt/partial/abandon, OR faulted 2xx → request_incomplete
-			  --   * everything else faulted → request_faulted
-			  SELECT toString(nr.ts) AS ts,
-			         multiIf(
-			           positionCaseInsensitive(fault_type, 'timeout') > 0, 'request_timeout',
-			           positionCaseInsensitive(fault_type, 'corrupt') > 0
-			             OR positionCaseInsensitive(fault_type, 'partial') > 0
-			             OR positionCaseInsensitive(fault_type, 'abandon') > 0
-			             OR (status >= 200 AND status < 300), 'request_incomplete',
-			           'request_faulted'
-			         ) AS type,
-			         concat(fault_type, ' ', method, ' ', path) AS info, 0
-			  FROM %s.network_requests AS nr WHERE %s AND faulted = 1
-			  UNION ALL
-			  -- Slow non-error requests: client_wait_ms > 2 s.
-			  SELECT toString(nr.ts) AS ts, 'slow_request' AS type,
-			         concat(toString(round(client_wait_ms, 0)), 'ms ', method, ' ', path) AS info, 0
-			  FROM %s.network_requests AS nr
-			  WHERE %s AND client_wait_ms > 2000 AND status < 400 AND faulted = 0
-			  UNION ALL
-			  -- Slow media-segment transfers: distinct from slow_request,
-			  -- this is about long actual transfer time on a segment-
-			  -- shaped URL. Mirrors the waterfall's slow-transfer class.
-			  SELECT toString(nr.ts) AS ts, 'slow_segment' AS type,
-			         concat(toString(round(transfer_ms, 0)), 'ms ', method, ' ', path) AS info, 0
-			  FROM %s.network_requests AS nr
-			  WHERE %s AND transfer_ms > 6000
-			    AND match(path, '\.(m4s|ts|mp4|m4a|m4v|aac|webm|mp3)($|\?)')
-			    AND status < 400 AND faulted = 0
-			  UNION ALL
-			  -- Retries: same URL fetched again within ~4 s. Detected
-			  -- via lagInFrame partitioned by url; the network log uses
-			  -- a per-segment-duration threshold but a fixed 4 s catches
-			  -- almost every real retry case (segments are 2 s/6 s).
-			  SELECT toString(retry_ts) AS ts, 'request_retry' AS type,
-			         concat(method, ' ', path) AS info, 0
-			  FROM (
-			    SELECT nr.ts AS retry_ts, method, url, path,
-			           lagInFrame(nr.ts, 1, nr.ts) OVER (PARTITION BY url ORDER BY nr.ts ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING) AS prev_url_ts
-			    FROM %s.network_requests AS nr
-			    WHERE %s
-			  )
-			  WHERE prev_url_ts != retry_ts
-			    AND dateDiff('millisecond', prev_url_ts, retry_ts) BETWEEN 1 AND 4000
-			)
-			-- DESC so the LIMIT keeps the *most recent* N events. ASC
-			-- caused rare events (HTTP 4xx fault injection, errors) to
-			-- be squeezed out of the result by high-volume ABR upshifts
-			-- / downshifts that dominate the oldest 500 rows on a long
-			-- session. Consumers sort client-side for display.
-			ORDER BY ts DESC
-			LIMIT %d
-			FORMAT JSONEachRow`,
-			// stall_or_buffer_pairs / rate_shifts / base CTEs
-			cfg.chDatabase, cfg.chTable, where,
-			cfg.chDatabase, cfg.chTable, where,
-			cfg.chDatabase, cfg.chTable, where,
-			// frozen/segment_stall, restart, playback_start, timejump, error (5)
-			cfg.chDatabase, cfg.chTable, where,
-			cfg.chDatabase, cfg.chTable, where,
-			cfg.chDatabase, cfg.chTable, where,
-			cfg.chDatabase, cfg.chTable, where,
-			cfg.chDatabase, cfg.chTable, where,
-			// user_marked + catch-all (2)
-			cfg.chDatabase, cfg.chTable, where,
-			cfg.chDatabase, cfg.chTable, where,
-			// HAR events: http_5xx, http_4xx, faulted, slow_request,
-			// slow_segment, request_retry — each filtered by harWhere
-			// (time range, NOT play_id) since iOS strips play_id from
-			// variant/segment URLs.
-			cfg.chDatabase, harWhere,
-			cfg.chDatabase, harWhere,
-			cfg.chDatabase, harWhere,
-			cfg.chDatabase, harWhere,
-			cfg.chDatabase, harWhere,
-			cfg.chDatabase, harWhere,
-			limit)
+		ep := eventsQueryParams{
+			SessionID: session,
+			PlayerID:  playerID,
+			PlayID:    r.URL.Query().Get("play_id"),
+			Limit:     limit,
+		}
+		query, params, err := buildEventsQuery(cfg, ep)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusBadRequest)
+			return
+		}
 		proxyClickHouseJSON(w, r, cfg, query, params)
 	}
+
 	mux.HandleFunc("/api/session_events", sessionEventsHandler)
 	mux.HandleFunc("/api/v2/session_events", sessionEventsHandler)
